[REFACTOR] Refactor return type Promo Code Utils

+++ src/main/java/id/ac/ui/cs/advprog/hoomgroompayment/util/PromoCodeUtil.java
@@ -1,9 +1,9 @@
 package id.ac.ui.cs.advprog.hoomgroompayment.util;
 public class PromoCodeUtil {
-    public static int extractDiscountPercentage(String promoCode) {
+    public static double extractDiscountPercentage(String promoCode) {
         String[] parts = promoCode.split("(?<=\\D)(?=\\d)");
         if (parts.length > 1) {
-            int discountPercentage = Integer.parseInt(parts[1]);
+            double discountPercentage = Double.parseDouble(parts[1]);
             return discountPercentage;
         } else {
             return 0;
